+++ utils.py
@@ -1,0 +1,20 @@
+
+
+
+def tag_list_to_dict(tags_list):
+    tags_dict = {}
+    for tag in tags_list:
+        if 'key' in tag and not tag['key'].startswith('aws:'):
+            tags_dict[tag['key']] = tag['value']
+        elif 'Key' in tag and not tag['Key'].startswith('aws:'):
+            tags_dict[tag['Key']] = tag['Value']
+
+    return tags_dict
+
+
+def tag_dict_to_list(tags_dict):
+    tags_list = []
+    for k, v in tags_dict.items():
+        tags_list.append({'Key': k, 'Value': v})
+
+    return tags_list
